wherewasi-queue-processor/requirements.txt
google-cloud-pubsub
python-dotenv
google-cloud-spanner